added archive function for notes (daily goal, profile)

+++ data/php/delete_file_from_download.php
@@ -1,15 +1,34 @@
 <?php
-    if(isset($_POST['file'])){
+    if(isset($_POST['file']) && isset($_POST['directory'])){
         $filename = $_POST['file'];
         $nickname = $_POST['nickname'];
 
-        $dir=$_SERVER['DOCUMENT_ROOT']."/tmp/task_table/";
+        $directory = $_POST['directory'];
+
+        if($directory == ""){
+            $dir = $_SERVER['DOCUMENT_ROOT'] . "/tmp/task_table/";
+        }
+        else{
+            $dir = $_SERVER['DOCUMENT_ROOT'] . $directory;
+        }
         if(is_dir($dir)){
-            if($dir_num=opendir($dir)){
-                while($file=readdir($dir_num)){
+            if($dir_num = opendir($dir)){
+                while($file = readdir($dir_num)){
                     try{
-                        $file_user = explode("_",$file)[0];
-                        if($nickname == $file_user){
+                        if($directory == ""){
+                            $file_user = explode("_",$file)[0];
+                            if($nickname == $file_user){
+                                if($file == $filename){
+                                    if (!unlink($dir.$filename)) {
+                                        echo ("Не удалось");
+                                    }
+                                    else {
+                                        echo ("Успешно");
+                                    }
+                                }
+                            }
+                        }
+                        else{
                             if($file == $filename){
                                 if (!unlink($dir.$filename)) {
                                     echo ("Не удалось");
